apps/users/api_endpoints/panel/Login/views.py
from rest_framework.generics import GenericAPIView
from rest_framework.permissions import AllowAny
from rest_framework_simplejwt.tokens import RefreshToken
from rest_framework.response import Response


from .serializers import LoginSerializer


class LoginView(GenericAPIView):
    serializer_class = LoginSerializer
    permission_classes = (AllowAny,)

    def post(self, request, *args, **kwargs):
        serializer = self.get_serializer(data=request.data, context={'request': request})
        serializer.is_valid(raise_exception=True)
        user = serializer.validated_data
        tokens = RefreshToken.for_user(user)

        return Response({"access_token": str(tokens.access_token), "refresh_token": str(tokens)})


__all__ = [
    'LoginView',
]
