requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-modeltranslation
django-redis
nplusone==1.0.0
celery==5.3.6
redis==5.0.1
flower==2.0.1
dj-database-url==2.2.0
djangorestframework-simplejwt==4.8.0
